refactor(federation): barrels
